SaLT version is not hardcoded anymore.

+++ .getsalt.sh
@@ -3,8 +3,12 @@
 # Used to download the correct version of SaLT.
 
 cd "$(dirname "$0")"
-SALT_VER='master'
-SALT_URL='git://github.com/jrd/SaLT.git'
+SALT_VER="$1"
+if [ -z "$SALT_VER" ]; then
+  echo "You need to specify a branch or tag for SaLT" >&2
+  exit 1
+fi
+SALT_URL='git://github.com/Salix-OS/SaLT.git'
 if [ -d salt ]; then
   rm -rf salt || echo "salt directory cannot be removed, check permissions" >&2
 fi
